trading-backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic-settings==2.1.0

# Binance API
python-binance==1.0.19
ccxt==4.2.25

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0  # Async PostgreSQL driver (better for FastAPI)
alembic==1.13.1

# AI/ML Libraries
openai==1.10.0
anthropic==0.8.1
transformers==4.37.0
torch>=2.6.0  # PyTorch for ML models
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
ta==0.11.0  # Technical analysis library

# WebSocket
websockets==12.0
python-socketio==5.11.0

# Security
cryptography==42.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# Environment variables
python-dotenv==1.0.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0  # For Telegram Bot API

# Background tasks
celery==5.3.6
celery-types==0.20.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
